move cryption functions to utility
